fix files reorganized
